+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### [1.5.9]
 
 * Add command to change chat names
+* Add init scripts example
 
 #### [1.5.8]