+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-#### [1.5.9]
+#### [1.6.0]
 
 * Replace ruby wrapper for Telegram's Bot API telegrammer to telegram-bot-ruby
+* Change command /task
+
+#### [1.5.9]
+
 * Add to send current value for group commands
 * Add command to change chat names
 * Add init scripts example